+++ cloudmodel/cloudmodel.go
@@ -100,13 +100,3 @@ func serving(t *Traits, w http.ResponseWriter, r *http.Request, servicePath stri
 		http.Error(w, "Invalid service request [Do not modify the services subpath in the configuration file]", http.StatusBadRequest)
 	}
 }
-
-// aggregate handles GET requests to generate the merged cloud SysML v2 model.
-func (t *Traits) aggregate(w http.ResponseWriter, r *http.Request) {
-	switch r.Method {
-	case "GET":
-		t.assembleModel(w)
-	default:
-		http.Error(w, "Method is not supported.", http.StatusNotFound)
-	}
-}
