+++ ds18b20/thing.go
@@ -18,6 +18,7 @@ package main
 
 import (
 	"context"
+	"fmt"
 	"log"
 	"net/http"
 	"os"
@@ -96,6 +97,36 @@ func newResource(configuredAsset usecases.ConfigurableAsset, sys *components.Sys
 
 	return ua, func() {
 		log.Printf("disconnecting from %s\n", configuredAsset.Name)
+	}
+}
+
+//-------------------------------------Service handlers
+
+// readTemp gets the unit asset's temperature datum and sends it in a signal form
+func (t *Traits) readTemp(w http.ResponseWriter, r *http.Request) {
+	switch r.Method {
+	case "GET":
+		getMeasuremet := STray{
+			Action: "read",
+			ValueP: make(chan forms.SignalA_v1a),
+			Error:  make(chan error),
+		}
+		t.trayChan <- getMeasuremet
+		select {
+		case err := <-getMeasuremet.Error:
+			fmt.Printf("Logic error in getting measurement, %s\n", err)
+			w.WriteHeader(http.StatusInternalServerError) // Use 500 for an internal error
+			return
+		case temperatureForm := <-getMeasuremet.ValueP:
+			usecases.HTTPProcessGetRequest(w, r, &temperatureForm)
+			return
+		case <-time.After(5 * time.Second): // Optional timeout
+			http.Error(w, "Request timed out", http.StatusGatewayTimeout)
+			log.Println("Failure to process temperature reading request")
+			return
+		}
+	default:
+		http.Error(w, "Method is not supported.", http.StatusNotFound)
 	}
 }
