+++ kgrapher/kgrapher.go
@@ -107,25 +107,3 @@ func serving(t *Traits, w http.ResponseWriter, r *http.Request, servicePath stri
 	}
 }
 
-// aggregate writes out the knowledge graph of the local cloud and pushes it to GraphDB
-func (t *Traits) aggregate(w http.ResponseWriter, r *http.Request) {
-	switch r.Method {
-	case "GET":
-		t.assembleOntologies(w)
-	default:
-		http.Error(w, "Method is not supported.", http.StatusNotFound)
-	}
-}
-
-// listOntologies writes out the HTML produced by localOntologies()
-func (t *Traits) listOntologies(w http.ResponseWriter, r *http.Request) {
-	switch r.Method {
-	case "GET":
-		p := r.Pattern
-		html := t.localOntologies(p)
-		w.Header().Set("Content-Type", "text/html; charset=utf-8")
-		fmt.Fprint(w, html)
-	default:
-		http.Error(w, "Method is not supported.", http.StatusNotFound)
-	}
-}
