+++ messenger/thing.go
@@ -12,6 +12,7 @@ import (
 	"strconv"
 	"strings"
 	"sync"
+	"testing"
 	"time"
 
 	"github.com/sdoque/mbaigo/components"
@@ -94,6 +95,79 @@ func newResource(ca usecases.ConfigurableAsset, sys *components.System) (*compon
 	go t.runBeacon()
 	f := func() {}
 	return ua, f, nil
+}
+
+//-------------------------------------Service handlers
+
+func (t *Traits) handleNewMessage(w http.ResponseWriter, r *http.Request) {
+	if r.Method != http.MethodPost {
+		http.Error(w, http.StatusText(http.StatusMethodNotAllowed), http.StatusMethodNotAllowed)
+		return
+	}
+	body, err := io.ReadAll(r.Body)
+	if err != nil {
+		http.Error(w, http.StatusText(http.StatusInternalServerError),
+			http.StatusInternalServerError)
+		return
+	}
+	defer r.Body.Close()
+
+	form, err := usecases.Unpack(body, r.Header.Get("Content-Type"))
+	if err != nil {
+		http.Error(w, http.StatusText(http.StatusBadRequest), http.StatusBadRequest)
+		return
+	}
+	msg, ok := form.(*forms.SystemMessage_v1)
+	if !ok {
+		http.Error(w, http.StatusText(http.StatusBadRequest), http.StatusBadRequest)
+		return
+	}
+	t.addMessage(*msg) // Don't want to have to deal with pointers, hence the *
+}
+
+// Encapsulates the regular bytes.Buffer, in order to allow causing mock errors
+type mockableBuffer struct {
+	bytes.Buffer // This embedded struct is available as "mockableBuffer.Buffer" by default
+	errWrite     error
+}
+
+func (mock *mockableBuffer) setWriteError(err error) {
+	mock.errWrite = err
+}
+
+func (mock *mockableBuffer) Write(body []byte) (int, error) {
+	if mock.errWrite != nil {
+		return 0, mock.errWrite
+	}
+	return mock.Buffer.Write(body)
+}
+
+const testBufferHeader string = "x-testing-buffer"
+
+func (t *Traits) handleDashboard(w http.ResponseWriter, r *http.Request) {
+	if r.Method != http.MethodGet {
+		http.Error(w, http.StatusText(http.StatusMethodNotAllowed), http.StatusMethodNotAllowed)
+		return
+	}
+	errors, warnings, latest := t.filterLogs()
+	data := map[string]any{
+		"Errors":   errors,
+		"Warnings": warnings,
+		"Latest":   latest,
+	}
+
+	buf := &mockableBuffer{}
+	// Protects the special test header by enabling it's use only while running `go test`
+	if testing.Testing() && r.Header.Get(testBufferHeader) != "" {
+		// This write error will cause an error in the template.Execute() below
+		buf.setWriteError(fmt.Errorf("mock error"))
+	}
+	if err := t.tmplDashboard.Execute(buf, data); err != nil {
+		usecases.LogError(t.owner, "execute dashboard: %s", err)
+		http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
+		return
+	}
+	buf.WriteTo(w) // Ignoring errors, can't do much with them anyways if the transfer fails
 }
 
 ////////////////////////////////////////////////////////////////////////////////
