+++ photographer/photographer.go
@@ -103,17 +103,3 @@ func serving(t *Traits, w http.ResponseWriter, r *http.Request, servicePath stri
 	}
 }
 
-func (t *Traits) photograph(w http.ResponseWriter, r *http.Request) {
-	switch r.Method {
-	case "GET":
-		fileForm, err := t.takePicture()
-		if err != nil {
-			log.Println(err)
-			http.Error(w, "Failed to take a picture, check if PiCam is connected", http.StatusNotFound)
-			return
-		}
-		usecases.HTTPProcessGetRequest(w, r, &fileForm)
-	default:
-		http.Error(w, "Method is not supported.", http.StatusNotFound)
-	}
-}
