+++ sailor/sailor.go
@@ -37,7 +37,6 @@ import (
 	"time"
 
 	"github.com/sdoque/mbaigo/components"
-	"github.com/sdoque/mbaigo/forms"
 	"github.com/sdoque/mbaigo/usecases"
 )
 
@@ -104,27 +103,3 @@ func serving(t *Traits, w http.ResponseWriter, r *http.Request, servicePath stri
 	}
 }
 
-// access handles GET requests for the signal's current value.
-// It sends an STray to the signal's assetLoop and blocks until a reply arrives.
-func (t *Traits) access(w http.ResponseWriter, r *http.Request) {
-	switch r.Method {
-	case http.MethodGet:
-		order := STray{
-			ValueP: make(chan forms.SignalA_v1a),
-			Error:  make(chan error),
-		}
-		t.trayChan <- order
-		select {
-		case err := <-order.Error:
-			log.Printf("access %s: error: %v", t.name, err)
-			http.Error(w, "internal error", http.StatusInternalServerError)
-		case f := <-order.ValueP:
-			usecases.HTTPProcessGetRequest(w, r, &f)
-		case <-time.After(5 * time.Second):
-			http.Error(w, "request timed out", http.StatusGatewayTimeout)
-			log.Printf("access %s: timeout", t.name)
-		}
-	default:
-		http.Error(w, "method not supported", http.StatusMethodNotAllowed)
-	}
-}
