+++ thermostat/thermostat.go
@@ -107,43 +107,3 @@ func serving(t *Traits, w http.ResponseWriter, r *http.Request, servicePath stri
 	}
 }
 
-// setpt handles the get and set requests for the thermostat set point
-func (t *Traits) setpt(w http.ResponseWriter, r *http.Request) {
-	switch r.Method {
-	case "GET":
-		setPointForm := t.getSetPoint()
-		usecases.HTTPProcessGetRequest(w, r, &setPointForm)
-	case "PUT":
-		sig, err := usecases.HTTPProcessSetRequest(w, r)
-		if err != nil {
-			log.Println("Error with the setting request of the position ", err)
-		}
-		t.setSetPoint(sig)
-		confirmed := t.getSetPoint()
-		usecases.HTTPProcessGetRequest(w, r, &confirmed)
-	default:
-		http.Error(w, "Method is not supported.", http.StatusNotFound)
-	}
-}
-
-// diff handles the get requests for the thermostat error signal
-func (t *Traits) diff(w http.ResponseWriter, r *http.Request) {
-	switch r.Method {
-	case "GET":
-		signalErr := t.getError()
-		usecases.HTTPProcessGetRequest(w, r, &signalErr)
-	default:
-		http.Error(w, "Method is not supported.", http.StatusNotFound)
-	}
-}
-
-// variations handles the get requests for the thermostat jitter signal
-func (t *Traits) variations(w http.ResponseWriter, r *http.Request) {
-	switch r.Method {
-	case "GET":
-		signalErr := t.getJitter()
-		usecases.HTTPProcessGetRequest(w, r, &signalErr)
-	default:
-		http.Error(w, "Method is not supported.", http.StatusNotFound)
-	}
-}
