+++ tracker/thing.go
@@ -19,10 +19,13 @@ package main
 import (
 	"database/sql"
 	"fmt"
+	"io"
 	"log"
+	"mime"
 	"net/http"
 	"os"
 	"reflect"
+	"strconv"
 	"time"
 
 	"github.com/sdoque/mbaigo/components"
@@ -136,6 +139,99 @@ func (f *PenHolderOrder_v1) FormVersion() string {
 
 func init() {
 	forms.FormTypeMap["PenHolderOrder_v1"] = reflect.TypeOf(PenHolderOrder_v1{})
+}
+
+//-------------------------------------Service handlers
+
+// orderHandler handles GET (retrieve by ?id=N), POST (new order), and PUT (update order).
+func (t *Traits) orderHandler(w http.ResponseWriter, r *http.Request) {
+	switch r.Method {
+
+	case http.MethodGet:
+		idStr := r.URL.Query().Get("id")
+		email := r.URL.Query().Get("email")
+		if idStr == "" || email == "" {
+			http.Error(w, "missing query parameters: id and email are both required", http.StatusBadRequest)
+			return
+		}
+		id, err := strconv.Atoi(idStr)
+		if err != nil {
+			http.Error(w, "id must be an integer", http.StatusBadRequest)
+			return
+		}
+		order, err := GetOrderByIDAndEmail(t.db, id, email)
+		if err != nil {
+			http.Error(w, err.Error(), http.StatusNotFound)
+			return
+		}
+		order.NewForm()
+		usecases.HTTPProcessGetRequest(w, r, order)
+
+	case http.MethodPost, http.MethodPut:
+		contentType := r.Header.Get("Content-Type")
+		mediaType, _, err := mime.ParseMediaType(contentType)
+		if err != nil {
+			http.Error(w, "could not parse Content-Type: "+err.Error(), http.StatusBadRequest)
+			return
+		}
+		defer r.Body.Close()
+		bodyBytes, err := io.ReadAll(r.Body)
+		if err != nil {
+			http.Error(w, "reading request body: "+err.Error(), http.StatusBadRequest)
+			return
+		}
+
+		unpacked, err := usecases.Unpack(bodyBytes, mediaType)
+		if err != nil {
+			http.Error(w, "unpacking request: "+err.Error(), http.StatusBadRequest)
+			return
+		}
+		record, ok := unpacked.(*PenHolderOrder_v1)
+		if !ok {
+			http.Error(w, "expected PenHolderOrder_v1 body", http.StatusBadRequest)
+			return
+		}
+
+		if record.OrderNumber <= 0 {
+			// New order: insert and forward to downstream system.
+			oN, err := InsertOrder(t.db, record)
+			if err != nil {
+				http.Error(w, "inserting order: "+err.Error(), http.StatusInternalServerError)
+				return
+			}
+			record.OrderNumber = oN
+			log.Printf("tracker: new order %d filed\n", oN)
+
+			// Forward the new order to the addorder cervice (e.g. a production TSP).
+			if t.owner != nil {
+				packed, err := usecases.Pack(record, "application/json")
+				if err == nil {
+					if _, err := usecases.SetState(t.cervices["addorder"], t.owner, packed); err != nil {
+						log.Printf("tracker: could not forward order %d to addorder service: %v\n", oN, err)
+					}
+				}
+			}
+		} else {
+			// Existing order: update in place.
+			if err := UpdateOrder(t.db, record); err != nil {
+				http.Error(w, "updating order: "+err.Error(), http.StatusInternalServerError)
+				return
+			}
+			log.Printf("tracker: order %d updated\n", record.OrderNumber)
+		}
+
+		confirmed, err := usecases.Pack(record, mediaType)
+		if err != nil {
+			http.Error(w, "marshalling response: "+err.Error(), http.StatusInternalServerError)
+			return
+		}
+		w.Header().Set("Content-Type", mediaType)
+		w.WriteHeader(http.StatusOK)
+		w.Write(confirmed) //nolint:errcheck
+
+	default:
+		http.Error(w, "Method is not supported.", http.StatusMethodNotAllowed)
+	}
 }
 
 //-------------------------------------Database
